Implemented functionality to display details about the moon and all dwarf planets.
