+++ include/menu.h
@@ -2,18 +2,16 @@
 #define MENU_H
 
 #include "planet.h"
+#include "moon.h"
 
 void printGreeting();
 void printGoodbye();
-void printPlanetMenu();
-void printMoonMenu();
 void printDwarfPlanetMenu();
 void printSunMenu();
 void printAsteroidMenu();
 void printCometMenu();
 void printMeteorMenu();
 void printTelescopeMenu();
-void printPlanetData(struct Planet planet);
 void printCalculationMenu(struct Planet planet);
 void printPlanetaryCalculator();
 void printRelativeSizeCalcMenu();
