+++ src/moon.c
@@ -2,3 +2,12 @@
 #include "../include/moon.h"
 
 // define functions in this file
+
+void printMoonMenu() {
+    printf("\nThe Moon\n");
+    printf("----------------\n\n");
+    printf("The Moon makes Earth more livable by moderating our home planet's wobble on its axis, leading to a relatively stable climate. ");
+    printf("It also causes tides, creating a rhythm that has guided humans for thousands of years.\n\n");
+    printf("The Moon rotates exactly once each time it orbits our planet.\n\n");
+    printf("Observers on Earth can track the Moon's motion in space (relative to Earth and the Sun) through moon phases, supermoons and eclipses.\n\n");
+}
